0.13.0

+++ external/flatcc/flatcc.h
@@ -16,6 +16,8 @@ extern "C" {
  * which has a life of its own.
  */
 
+#include <stddef.h>
+
 #ifndef UINT8_MAX
 #include <stdint.h>
 #endif
@@ -26,7 +28,7 @@ extern "C" {
 #endif
 
 typedef struct flatcc_options flatcc_options_t;
-typedef void (*flatcc_error_fun) (void *err_ctx, const char *buf, int len);
+typedef void (*flatcc_error_fun) (void *err_ctx, const char *buf, size_t len);
 
 struct flatcc_options {
     size_t max_schema_size;
